Refactor CLI into package
